resource: remove symbol table
Handshake resource data structures included two forms of compression
to reuse redundant strings: A symbol table implemented in hsd, and
"legacy" DNS Message Compression defined in rfc1035 section 4.1.4

The mainnet resource serialization no longer uses the symbol table.
In hnsd, the symbol table struct *also* included a copy of the full
DNS message so that compression pointers could be "looked up" when
used by the DNS compression protocol.

We still need to pass around the "DNS dump" to continue using that
protocol, but do away with the symbol table it was contained in.

For reference, use of the deprecated hsd symbol table was marked by
a 0x80 byte. Usage of the DNS compression scheme is marked by a byte
with the first two bits set (i.e. `char & 0xc0`)