resource: if resource has HSK NS, create DNS NS record
This step applies to all NS, GLUE and SYNTH records, and so a new
function hsk_resource_has_ns() is added to search the resource.

These records have explicit names and inet4/inet6 IPs, so
target_to_dns() can be substituted and the logic simplified
inside hsk_resource_to_ns() itself.

Finally the API for ip_to_b32() is changed to expect a raw byte array
(an IPv4 or IPv6 address) instead of an hsk_target_t. This disrupts
the current implementation of target_to_dns() but that function and
the type hsk_target_t will be removed completely in a future commit.