dns: refactor name_parse(): read wire format and keep in wire format
This function used to convert wire-formatted byte arrays to strings.
We are no longer using strings for names internally so this function
got much simpler. All it really needs to do now is un-compress a name
from a DNS message packet.

Because this function is the main entry point for all names in the
program, it is also critical to verify label and name size rules.
Once names are ingested by name_parse() and stored in structs, we
can for the most part assume they are valid.