dns: refactor name_serialize(): read and keep in wire format
This function used to convert names from strings to wire-formatted
byte arrays. We are no longer using strings for names internally so
this function got much simpler. All it really needs to do now is
compress a name into a DNS message by adding label pointers where
necessary.

Because this function is the main output point for names it is
critical to verify label and name lengths when we write.