dns: use dns-name-based hash map for label compression in serialize()
hsk_dns_name_serialize() takes an argument hsk_dns_cmp_t which
includes a generic map object hsk_map_t. This map is used to track
names and labels as a message is being written, so compression
pointers can be applied. Since names are byte arrays now and not
strings this map needs to be reimplemented with corrected hash and
equality functions.