dns: repurpose name_verify() to match hsd rules.verifyName()
In addition to procesisng byte arrays instead of strings, repurpose
this function to check a single label for validity as a Handshake TLD.

The most important time to use it is when receiving a user's request.
TLDs need to be valid before hashing and requesting Urkel proofs from
full nodes.